requirements.txt
accessible-pygments==0.0.5
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==0.7.16
anndata==0.11.4
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.9.0
archs4py==1.0.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
biomart==0.9.2
biopython==1.85
biothings_client==0.4.1
bleach==6.2.0
blinker==1.9.0
botocore==1.37.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
clustergrammer2==0.18.0
clustergrammer_widget==1.13.3
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
dash_bio==1.0.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docutils==0.21.2
entrypoints==0.4
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.58.2
formulaic==1.1.1
formulaic-contrasts==1.0.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
geofetch==0.12.7
GEOparse==2.0.4
google-auth==2.40.3
google-genai==1.24.0
greenlet==3.2.3
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.3.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==7.8.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-book==1.0.4.post1
jupyter-cache==1.0.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.8.1
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.11
kaleido==1.0.0
kiwisolver==1.4.8
latexcodec==3.0.1
linkify-it-py==2.0.3
llvmlite==0.44.0
loess==2.1.2
logistro==1.1.0
logmuse==0.2.8
lxml==5.4.0
maayanlab-bioinformatics @ git+https://github.com/Maayanlab/maayanlab-bioinformatics.git@5b38cf2ce8f67928777852b69f2e2659c6eb9043
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.1
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.4
mygene==3.2.2
myst-nb==1.3.0
myst-parser==3.0.1
narwhals==1.42.1
natsort==8.4.0
nbclassic==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==6.1.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas
pandocfilters==1.5.1
papermill==2.6.0
ParmEd==4.3.0
parso==0.8.4
patsy==1.0.1
pephubclient==0.4.5
peppy==0.40.7
periodictable==2.0.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotbin==3.1.7
plotly==6.1.2
progress==1.6
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybtex==0.25.1
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydata-sphinx-theme==0.15.4
pydeseq2==0.5.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pysradb==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qnorm==0.9.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-ftp==0.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rpy2==3.6.1
rpy2-rinterface==3.6.1
rpy2-robjects==3.6.1
rsa==4.9.1
s3fs==2025.5.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
session_info==1.0.1
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==7.4.7
sphinx-book-theme==1.1.4
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==1.0.0
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.3.1
sphinx-togglebutton==0.3.2
sphinx_design==0.6.1
sphinx_external_toc==1.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ubiquerg==0.8.1
uc-micro-py==1.0.3
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wget==3.2
wheel==0.45.1
widgetsnbextension==3.6.10
wrapt==1.17.2
xalign @ git+https://github.com/MaayanLab/xalign.git@9b4a33636045e26838bb91b31965ba44b2db9d3b
xlrd==2.0.2
xmltodict==0.14.2
yarl==1.20.1
zipp==3.23.0
